requirements.txt
flask
python-binance
python-dotenv
requests
pandas
numpy
waitress
pydantic
pydantic-settings